+++ packages/adapter-interfaces/src/adapters/Sessions.ts
@@ -1,7 +1,7 @@
-import { Session } from "../types";
+import { Session, SessionInsert } from "../types";
 
 export interface SessionsAdapter {
-  create: (session: Session) => Promise<void>;
+  create: (tenant_id: string, session: SessionInsert) => Promise<Session>;
   get: (tenant_id: string, id: string) => Promise<Session | null>;
   update: (
     tenant_id: string,