+++ packages/drizzle/src/index.ts
@@ -1,4 +1,4 @@
-import { tmp, ApplicationInsert } from "@authhero/adapter-interfaces";
+import { tmp } from "@authhero/adapter-interfaces";
 
 export default function create() {
   console.log(tmp);